+++ compiler/CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed typo on `Teams.derelict` (was `Teams.delerict`).
 - Fixed the internal implementations of `Math.sin`, `Math.cos` and `Math.tan`.
 - Fixed logical expressions creating unnecessary temp values.
 - Fixed optional chaning not working on null literals (`null?.foo`).